+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Changed
 - All plots now have a toolbox by default.
+### Added
+- histogram function.
+- fastcat sequence summary component.
 - Added ecStat 1.2.0 to vendor resources
 - Adding paging options to DataTable
 - CLI now configures logging.
-
 
 ## [v0.2.2]
 ### Added